refactor(#8/post): exception 클래스 위치 이동
